deep-signal-radar-core/requirements.txt
feedparser==6.0.11
requests==2.32.3
google-generativeai==0.8.3
openai==1.55.3
pydantic==2.10.3
python-dotenv==1.0.1
